+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Mounts `/libretro/core` as the directory where the core was loaded from
 - Adds `love.filesystem.getSaveDirectory()`
 - Adds `love.filesystem.newFileData(contents, name)`
+- Adds `love.filesystem.getExecutablePath()`
 
 ### Fixes
 - Fixed loading ChaiLove without active content